api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0.0
requests>=2.31.0
elasticsearch>=7.17.0,<8.0.0
pytest>=8.0.0
pytest-django>=4.8.0
flake8>=6.1.0
ruff==0.11.0
kombu>=5.3.4,<6.0

